Cleaned up a couple things and also fixed a couple typos on the cooldown+bloom implementation. Changed Bloom from a float to a double because it's always between 0 and 1 I literally do not floating point. KNOWN ISSUE: First person arm animations are weird. not sure how to fix that yet

+++ src/main/java/net/moppzarella/mzguns/component/ModDataComponents.java
@@ -19,8 +19,8 @@ public class ModDataComponents {
 
    public static final DeferredHolder<DataComponentType<?>, DataComponentType<Integer>> CURRENT_FIRING_COOLDOWN = register("current_firing_cooldown",
             builder -> builder.persistent(Codec.INT));
-    public static final DeferredHolder<DataComponentType<?>, DataComponentType<Float>> CURRENT_BLOOM = register("current_bloom",
-            builder -> builder.persistent(Codec.FLOAT ));
+    public static final DeferredHolder<DataComponentType<?>, DataComponentType<Double>> CURRENT_BLOOM = register("current_bloom",
+            builder -> builder.persistent(Codec.DOUBLE ));
 
     private static <T>DeferredHolder<DataComponentType<?>, DataComponentType<T>> register(String name,
                                                                                           UnaryOperator<DataComponentType.Builder<T>> builderOperator) {
